+++ sample/src/main/java/com/synerise/sdk/sample/ui/dev/apiAdapter/SyneriseSdkApi.java
@@ -28,6 +28,7 @@ public enum SyneriseSdkApi {
     REQUEST_EMAIL_CHANGE(R.string.client_request_email_change),
     CONFIRM_EMAIL_CHANGE(R.string.client_confirm_email_change),
     REQUEST_EMAIL_CHANGE_FACEBOOK(R.string.client_request_email_change_facebook),
+    RECOGNIZE_ANONYMOUS(R.string.client_reconize_anonymus),
 
     // PROMOTIONS API
     CLIENT_GET_PROMOTIONS(R.string.client_get_promotions),
@@ -69,7 +70,8 @@ public enum SyneriseSdkApi {
                 CONFIRM_PHONE_UPDATE,
                 REQUEST_EMAIL_CHANGE,
                 CONFIRM_EMAIL_CHANGE,
-                REQUEST_EMAIL_CHANGE_FACEBOOK);
+                REQUEST_EMAIL_CHANGE_FACEBOOK,
+                RECOGNIZE_ANONYMOUS);
     }
 
     public static List<SyneriseSdkApi> getPromotionApis() {